kishan_sathi_backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
python-decouple==3.8
requests==2.32.3
Pillow==11.1.0
cloudinary==1.44.0
django-cloudinary-storage==0.3.0
psycopg2-binary==2.9.10
firebase-admin==6.6.0
gunicorn==23.0.0
whitenoise==6.8.2
